Refactor ModelEvaluator to identify variables itself
This removes an API design wart: if someone passed only some of the
data's variables to ModelEvaluator, the bottom-model's uncertainty and
number of degrees of freedom would be silently wrong.

As a nice bonus, this reduces the asymptotic complexity of this function
from O(n*v) to O(n+v), where n is number of observed states and v is
number of variables.