Use less memory in Model::more_complex
My previous implementation had to accumulate all subsets of all
relations in the model. This version only stores data proportional to
the output length.

I haven't profiled any of this though, so who knows, it might actually
be worse. I'm committing it anyway because 1) I thought really hard
about the problem and want to record some history of things I tried, and
2) I think this version provides more insights to a careful reader about
ways it might be improved.