Test larger lattices in integration testing
This test needs to take a reasonable amount of time to walk the whole
lattice. At five variables, it finishes pretty quickly. At six variables
I don't have the patience to wait for it to finish even once, let alone
as a regular part of the test suite. Combinatorial explosions kinda suck
that way.

That said, I found bugs by going up to five variables instead of four,
so it's worth testing with as many variables as possible.