Make 3-relation case look like 2-relation case
I find it easier to read this way.

This does mean walking through the full A and B relations for each
choice of C, instead of looking at the already analyzed results, which
will be smaller if there were any variables in common. On the other hand
this way only iterates each C once instead of four times. So any
performance trade-off is probably complicated. I haven't bothered
profiling it, at any rate.