Make Model::less_complex O(n) instead of O(n^2)
...with respect to the number of relations examined per model generated,
anyway. I haven't figured out how to do the complexity analysis based on
the number of variables in the model yet, for example.

Calling model.add_relation in a loop had to examine the whole model each
time. This version instead makes a single pass over the model, splitting
it into two parts: new relations can only be added in the later part and
only if they aren't subsumed by relations in the earlier part.

This only costs extra space proportional to the number of variables in
the target relation, which is small compared to the model as a whole.

The only disappointing thing is that this implementation is way harder
to understand than the previous one.